+++ src/Components/Notification/Components/index.ts
@@ -1,0 +1,2 @@
+export { default as Banner } from './Banner';
+export * from './NotifyMessage';