+++ src/Components/SpriteSheet/SpriteSheet.js
@@ -141,6 +141,11 @@ const SpriteSheet = () => (
         <path d="M12 2C6.48 2 2 6.48 2 12s4.48 10 10 10 10-4.48 10-10S17.52 2 12 2zm1 15h-2v-2h2v2zm0-4h-2V7h2v6z" />
       </svg>
     </g>
+    <g id={ICON_NAMES.UPGRADE}>
+      <svg viewBox="6 2 16 18" fill="currentColor">
+        <path d="M16 18v2H8v-2h8zM11 7.99V16h2V7.99h3L12 4 8 7.99h3z"></path>
+      </svg>
+    </g>
     <g id={ICON_NAMES.HELP_OUTLINE}>
       <svg viewBox="0 0 24 24" fill="currentColor">
         <path d="M11 15h2v2h-2zm0-8h2v6h-2zm.99-5C6.47 2 2 6.48 2 12s4.47 10 9.99 10C17.52 22 22 17.52 22 12S17.52 2 11.99 2zM12 20c-4.42 0-8-3.58-8-8s3.58-8 8-8 8 3.58 8 8-3.58 8-8 8z" />