+++ src/Pages/Announcements/Components/index.ts
@@ -1,0 +1,1 @@
+export { default as AnnouncementsList } from './AnnouncementsList';